+++ src/exo1 - Basic types/exo1.exercise.ts
