+++ src/exo2 - Combinators/exo2.exercise.ts
