+++ src/exo3 - Sort with Ord/exo3.test.ts
@@ -7,7 +7,7 @@ import {
   sortPersonsByName,
   sortPersonsByAge,
   sortPersonsByAgeThenByName,
-} from './exo3.problem';
+} from './exo3.exercise';
 
 describe('exo3', () => {
   describe('sortStrings', () => {
