+++ src/exo5 - Nested data with traverse/exo5.exercise.ts
