+++ src/exo5 - Nested data with traverse/exo5.test.ts
@@ -7,7 +7,7 @@ import {
   performAsyncComputationInSequence,
   sequenceOptionArray,
   sequenceOptionTask,
-} from './exo5.problem';
+} from './exo5.exercise';
 
 describe('exo5', () => {
   describe('getCountryCurrencyOfOptionalCountryCode', () => {
