+++ src/exo6 - ReaderTaskEither/exo6.exercise.ts
