+++ src/exo8 - Own combinators/exo8.exercise.ts
