+++ src/exo8 - Own combinators/exo8.test.ts
@@ -7,7 +7,7 @@ import {
   bindEitherKW as rteBindEitherKW,
   bindReaderK as rteBindReaderK,
   bindReaderKW as rteBindReaderKW,
-} from './exo8.problem';
+} from './exo8.exercise';
 
 describe('exo8', () => {
   describe('bindEitherK[W]', () => {
